requirements.txt
discord_webhook
CloudFlare
python-dotenv
